+++ regression/reference/ownership/wrapownership.h
@@ -30,7 +30,7 @@ extern "C" {
 
 int * OWN_ReturnIntPtrRaw(void);
 
-int OWN_ReturnIntPtrScalar(void);
+int * OWN_ReturnIntPtrScalar(void);
 
 int OWN_ReturnIntPtrScalar_extrawrapper(void);