+++ regression/reference/pointers-cfi/wrappointers.cpp
@@ -1547,11 +1547,11 @@ void POI_returnIntPtrToFixedConstArray_CFI(CFI_cdesc_t *SHT_rv_cfi)
 // Attrs:     +deref(scalar)+intent(function)
 // Statement: c_function_native_*
 // start POI_returnIntScalar
-int POI_returnIntScalar(void)
+int * POI_returnIntScalar(void)
 {
     // splicer begin function.returnIntScalar
     int * SHC_rv = returnIntScalar();
-    return *SHC_rv;
+    return SHC_rv;
     // splicer end function.returnIntScalar
 }
 // end POI_returnIntScalar