+++ regression/reference/pointers-cfi/wrappointers.h
@@ -161,7 +161,7 @@ const int * POI_returnIntPtrToFixedConstArray(void);
 
 void POI_returnIntPtrToFixedConstArray_CFI(CFI_cdesc_t *SHT_rv_cfi);
 
-int POI_returnIntScalar(void);
+int * POI_returnIntScalar(void);
 
 int POI_returnIntScalar_extrawrapper(void);